requirements.txt
accelerate==0.34.2
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.0
asteroid-filterbanks==0.4.0
attrs==24.2.0
audioread==3.0.1
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clingo==5.7.1
colorlog==6.8.2
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
docopt==0.6.2
einops==0.8.0
eval_type_backport==0.2.0
fastapi==0.115.0
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.1
HyperPyYAML==1.2.2
idna==3.10
iniconfig==2.0.0
insanely-fast-whisper==0.0.15
Jinja2==3.1.4
joblib==1.4.2
jsonpath-python==1.0.6
julius==0.2.7
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.4.0
lightning-utilities==0.11.7
llvmlite==0.43.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mistralai==1.1.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.10.0.84
optuna==4.0.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
primePy==1.3
protobuf==5.28.2
psutil==6.0.0
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
PyAudio==0.2.14
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
pytest==8.3.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.12
python-sat==0.1.8.dev16
pytorch-lightning==2.4.0
pytorch-metric-learning==2.6.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.5.0.post1
speechbrain==1.0.1
SQLAlchemy==2.0.35
starlette==0.38.6
sympy==1.13.3
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tokenizers==0.20.0
torch==2.4.1
torch-audiomentations==0.11.1
torch_pitch_shift==1.2.5
torchaudio==2.4.1
torchmetrics==1.4.2
tqdm==4.66.5
transformers==4.45.1
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uv==0.4.18
uvicorn==0.31.0
yarl==1.13.1
